ASoC: codecs: rk730: fix and update some clock configurations
This patch removes the two uncommon sampling rates of 12000 and 24000Hz,
because they are not in the definition of SNDRV_PCM_RATE_8000_192000 in
the sound framework:
...
 #define SNDRV_PCM_RATE_11025 (1U<<2) /* 11025Hz */
 #define SNDRV_PCM_RATE_16000 (1U<<3) /* 16000Hz */
 #define SNDRV_PCM_RATE_22050 (1U<<4) /* 22050Hz */
 #define SNDRV_PCM_RATE_32000 (1U<<5) /* 32000Hz */
...

According to the configuration of the sampling rate group clock, the
configuration of 11025 and 22050Hz is added. And, it is changed to
select the appropriate mclk in the coeff_clk array according to the
sampling by the codec by default. If you need to use the previous
fixed FS method, you need to specify the property "rockchip,mclk-fs-fixed"
in the dts node.

Change-Id: Iae4478b807f4e6eb88bf6cc9f431ce2f725a5b36
Signed-off-by: Xing Zheng <zhengxing@rock-chips.com>